Update test refs
